Do not attemp delivery retries when testing endpoint
