+++ app/controllers/endpoints_controller.rb
@@ -43,9 +43,10 @@ class EndpointsController < ApplicationController
   end
 
   def test
-    job = SendNotificationJob.perform_now(@endpoint, Time.zone.now.to_s,
-                                          'Test', 'Test', 'Test',
-                                          'Test message from webhooks')
+    job = SendTestNotificationJob.perform_now(@endpoint, Time.zone.now.to_s,
+                                              'Test', 'Test', 'Test',
+                                              'Test message from webhooks')
+
     if job.nil?
       head :no_content
       return
